More work on splitup
