+++ api-model/src/main/java/io/enmasse/iot/model/v1/InfinispanDeviceRegistry.java
@@ -5,8 +5,11 @@
 
 package io.enmasse.iot.model.v1;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonUnwrapped;
+
 import io.fabric8.kubernetes.api.model.Doneable;
 import io.sundr.builder.annotations.Buildable;
 import io.sundr.builder.annotations.Inline;
@@ -18,19 +21,28 @@ import io.sundr.builder.annotations.Inline;
         inline = @Inline(
                 type = Doneable.class,
                 prefix = "Doneable",
-                value = "done"
-                )
-        )
+                value = "done"))
 @JsonInclude(NON_NULL)
 public class InfinispanDeviceRegistry {
 
+    @JsonUnwrapped
+    private CommonDeviceRegistry commonDeviceRegistry;
     private ContainerConfig container;
     private JavaContainerOptions java;
     private InfinispanDeviceRegistryServer server;
 
+    public CommonDeviceRegistry getCommonDeviceRegistry() {
+        return commonDeviceRegistry;
+    }
+
+    public void setCommonDeviceRegistry(CommonDeviceRegistry commonDeviceRegistry) {
+        this.commonDeviceRegistry = commonDeviceRegistry;
+    }
+
     public ContainerConfig getContainer() {
         return container;
     }
+
     public void setContainer(ContainerConfig container) {
         this.container = container;
     }
@@ -38,6 +50,7 @@ public class InfinispanDeviceRegistry {
     public InfinispanDeviceRegistryServer getServer() {
         return server;
     }
+
     public void setServer(InfinispanDeviceRegistryServer server) {
         this.server = server;
     }
@@ -45,6 +58,7 @@ public class InfinispanDeviceRegistry {
     public void setJava(JavaContainerOptions java) {
         this.java = java;
     }
+
     public JavaContainerOptions getJava() {
         return java;
     }
